+++ src/Components/PaymentMethod/DiscountedPrice.js
@@ -4,9 +4,9 @@ import { store } from "./PaymentMethodContainer";
 
 export const DiscountedPrice = () => {
   const {
-    state: { updatedPrice, percentOff }
+    state: { updatedPrice, percentOff, currentPlan }
   } = useContext(store);
-  const { default_currency, default_locale } = Pelcro.site.read();
+  const { default_locale } = Pelcro.site.read();
 
   if (percentOff) {
     return (
@@ -14,7 +14,7 @@ export const DiscountedPrice = () => {
         (-{percentOff}){" "}
         {getFormattedPriceByLocal(
           updatedPrice,
-          default_currency,
+          currentPlan.currency,
           default_locale
         )}
       </span>